+++ changelog.md
@@ -9,6 +9,7 @@
 - Added per gamemode no player rotation options
 - Added Effect Speed slider to pastel / chroma colours
 - Added Unlock Main Demon Levels
+- Added Uncomplete Level
 - Fixed being able to complete levels when playing from a startpos with auto safe mode disabled
 - Improved Blur BG Performance and switched to using Blur API
 - Fixed White Square when using Fade In animation